requirements.txt
django
pillow
mysqlclient
reportlab
gunicorn
python-dotenv